+++ crates/ltk_ritobin/src/ast/node.rs
@@ -1,9 +1,9 @@
 mod object;
 mod property;
-mod r#struct;
+mod root_object;
 pub mod value;
 
 pub use object::*;
 pub use property::*;
-pub use r#struct::*;
+pub use root_object::*;
 pub use value::*;
